Add explicit casts for virtual channel IDs (#875)
The latest version of `stream_channel` made the type explicit on the
method signature whereas before it was implicit with an `as` cast inside
the method. Most clients wouldn't be impacted is but this package is
because of it's use of `implicit-casts: false`.

+++ lib/src/runner/remote_listener.dart
@@ -96,7 +96,7 @@ class RemoteListener {
           queue.rest.listen((message) {
             assert(message["type"] == "suiteChannel");
             SuiteChannelManager.current.connectIn(message['name'] as String,
-                channel.virtualChannel(message['id']));
+                channel.virtualChannel(message['id'] as int));
           });
 
           if ((message['asciiGlyphs'] as bool) ?? false) glyph.ascii = true;
@@ -207,7 +207,7 @@ class RemoteListener {
     testChannel.stream.listen((message) {
       assert(message['command'] == 'run');
       _runLiveTest(test.load(_suite, groups: groups),
-          channel.virtualChannel(message['channel']));
+          channel.virtualChannel(message['channel'] as int));
     });
 
     return {
